Code/requirements.txt
altgraph==0.17.4
packaging==24.1
pycryptodome==3.20.0
PyInputPlus==0.2.12
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
PySimpleValidate==0.2.12
pyspellchecker==0.8.1
stdiomask==0.0.6
termcolor==2.4.0
